requirements.txt
duckdb==0.8.1
numpy
fastapi
uvicorn
langchain
langchain_community
llama-index
llama_index==0.11.22
pydantic
schemas
openpyxl
xlrd
typing-extensions
llama_index.embeddings.huggingface
llama-index-core
ollama
llama-index-llms-langchain
numpy
pdfminer.six
PyMuPDF
pytesseract
Pillow
langchain_ollama
watchdog
python-dotenv
logger
pandas==2.0.3
pyarrow==12.0.0
unidecode